Try to split player service
